+++ src/test/shell/example-cli/src/main/shell/vars/constants.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-# shellcheck source=../config/completion.sh
-
-# example-cli version - single source of truth for release management
-declare -gr gr_example_cli_version=v0.1.0